# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wheel fortune doc API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/wheel": {
            "get": {
                "tags": [
                    "Колеса"
                ],
                "summary": "Получение списка колес",
                "operationId": "aa9999bc0fdb846ad80e86b55717f1a3",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Фильтрация по имени колеса",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтрация по статусу",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Поле для сортировки",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Порядок сортировки",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ОК",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Wheel"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": {
                            "role": "admin"
                        }
                    }
                ]
            },
            "post": {
                "tags": [
                    "Колеса"
                ],
                "summary": "Создание колеса",
                "operationId": "fa13e5506c95f358f9ba940e00713ead",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "count_sectors",
                                    "animation",
                                    "date_start",
                                    "date_end",
                                    "days_of_week"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Акция мая"
                                    },
                                    "count_sectors": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "animation": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "date_start": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-01"
                                    },
                                    "date_end": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-05-31"
                                    },
                                    "days_of_week": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Понедельник",
                                            "Среда"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ОК",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "Акция мая"
                                        },
                                        "count_sectors": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "Не активно"
                                        },
                                        "animation": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "date_start": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-05-01"
                                        },
                                        "date_end": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2025-05-31"
                                        },
                                        "days_of_week": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "Понедельник",
                                                "Среда"
                                            ]
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-07T12:31:41.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-07T12:39:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизованный доступ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": {
                            "role": "admin"
                        }
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Sector": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "prize_type": {
                        "type": "string",
                        "example": "material_thing"
                    },
                    "prize_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "probability": {
                        "type": "string",
                        "example": "15"
                    },
                    "wheel_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "prize": {
                        "$ref": "#/components/schemas/Prize"
                    }
                },
                "type": "object"
            },
            "Prize": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ручка синяя шариковая"
                    },
                    "count": {
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
...
            "Wheel": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Акция мая"
                    },
                    "count_sectors": {
                        "type": "integer",
                        "example": 5
                    },
                    "status": {
                        "type": "string",
                        "example": "Не активно"
                    },
                    "animation": {
                        "type": "boolean",
                        "example": true
                    },
                    "date_start": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-01"
                    },
                    "date_end": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-31"
                    },
                    "days_of_week": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Понедельник",
                            "Среда"
                        ]
                    },
                    "sectors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Sector"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}